% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{al_band_model}
\alias{al_band_model}
\title{Band arc length of the normal reference curve (C back-end)}
\usage{
al_band_model(sigma, a = 0.05, b = 0.95, nodes = 400L)
}
\arguments{
\item{sigma}{scale of the normal reference;}

\item{a, b}{the probability band;}

\item{nodes}{quadrature panels.}
}
\value{
the model band arc length.
}
\description{
Band arc length of the normal reference curve (C back-end)
}
