% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{al_moments}
\alias{al_moments}
\title{Exact mean, variance and support of the arc-length statistic (C back-end)}
\usage{
al_moments(n)
}
\arguments{
\item{n}{sample size.}
}
\value{
list with mean, var, support.
}
\description{
Exact mean, variance and support of the arc-length statistic (C back-end)
}
