% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{al_pvalue}
\alias{al_pvalue}
\title{Saddlepoint right-tail probability of the arc-length statistic (C back-end)}
\usage{
al_pvalue(s, n)
}
\arguments{
\item{s}{observed statistic value;}

\item{n}{sample size.}
}
\value{
the right-tail p-value.
}
\description{
Saddlepoint right-tail probability of the arc-length statistic (C back-end)
}
