% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{al_statistic}
\alias{al_statistic}
\title{Arc-length goodness-of-fit statistic (C back-end)}
\usage{
al_statistic(u)
}
\arguments{
\item{u}{numeric probability-integral transforms in [0,1].}
}
\value{
the scalar arc length of the probability-plot ogive.
}
\description{
Arc-length goodness-of-fit statistic (C back-end)
}
