% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{arc_generator}
\alias{arc_generator}
\alias{pgen}
\alias{dgen}
\alias{rgen}
\title{Arc-length generator: transform a bounded-support base into a new distribution}
\usage{
arc_generator(dens, lower, upper, ngrid = 4000L)

pgen(q, obj)

dgen(q, obj)

rgen(n, obj)
}
\arguments{
\item{dens}{base density function (vectorised).}

\item{lower, upper}{bounded support.}

\item{ngrid}{grid resolution.}

\item{q}{quantiles.}

\item{obj}{an "arcgen" object.}

\item{n}{draws.}
}
\value{
an object of class "arcgen".
}
\description{
G(x) = S_F^[a,x] / S_F^[a,b], the normalised cumulative arc length of the base cumulative
distribution function, whose density is proportional to sqrt(1 + f(x)^2). Requires bounded support.
}
\section{Functions}{
\itemize{
\item \code{pgen()}: cumulative distribution function of the generated law

\item \code{dgen()}: density of the generated law

\item \code{rgen()}: random generation from the generated law

}}
