% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{arccirc}
\alias{arccirc}
\alias{qd_arccirc}
\alias{Q_arccirc}
\alias{darccirc}
\alias{trigmom_arccirc}
\alias{rho_arccirc}
\alias{rarccirc}
\title{The smooth circular arc-length family}
\usage{
arccirc(c3 = 0, mu = 0)

qd_arccirc(u, obj)

Q_arccirc(u, obj)

darccirc(theta, obj)

trigmom_arccirc(p, obj, nodes = 4096L)

rho_arccirc(obj, nodes = 4096L)

rarccirc(n, obj)
}
\arguments{
\item{c3}{shape parameter, |c3| <= arcc_c3max(); c3 = 0 is the circular uniform.}

\item{mu}{mean direction in radians.}

\item{u}{probabilities in [0,1].}

\item{obj}{an "arccirc" object.}

\item{theta}{angles in radians.}

\item{p}{integer order.}

\item{nodes}{number of rectangle-rule nodes; the integrand is C^1 periodic, so convergence is
geometric and the default is ample.}

\item{n}{sample size.}
}
\value{
an object of class "arccirc".
}
\description{
The smooth circular arc-length family
}
\section{Functions}{
\itemize{
\item \code{qd_arccirc()}: quantile density q(u) of the underlying linear family

\item \code{Q_arccirc()}: quantile function Q(u) of the underlying linear family

\item \code{darccirc()}: circular density

\item \code{trigmom_arccirc()}: trigonometric moment of order p

\item \code{rho_arccirc()}: mean resultant length

\item \code{rarccirc()}: random generation

}}
