% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{arccirc_fr}
\alias{arccirc_fr}
\alias{qd_arccirc_fr}
\alias{trigmom_arccirc_fr}
\alias{Q_arccirc_fr}
\alias{darccirc_fr}
\alias{rarccirc_fr}
\title{Fejer-Riesz form of the circular arc-length family}
\usage{
arccirc_fr(p, mu = 0)

qd_arccirc_fr(u, obj)

trigmom_arccirc_fr(pord, obj, nodes = 4096L)

Q_arccirc_fr(u, obj, nodes = 4096L)

darccirc_fr(theta, obj, nodes = 4096L)

rarccirc_fr(n, obj, nodes = 4096L)
}
\arguments{
\item{p}{complex vector of polynomial coefficients.}

\item{mu}{mean direction.}

\item{u}{probabilities in [0,1].}

\item{obj}{an "arccirc_fr" object.}

\item{pord}{integer order.}

\item{nodes}{Simpson cells used to accumulate Q.}

\item{theta}{angles in radians.}

\item{n}{sample size.}
}
\value{
an object of class "arccirc_fr".
}
\description{
q(u) = |P(exp(2 pi i u))|^2 normalised to integrate to one. Every such q is non-negative and
infinitely differentiable around the circle by construction, so neither an admissibility condition
nor a smoothness constraint arises, and the coefficients are unrestricted.
}
\section{Functions}{
\itemize{
\item \code{qd_arccirc_fr()}: quantile density

\item \code{trigmom_arccirc_fr()}: trigonometric moment of order p

\item \code{Q_arccirc_fr()}: quantile function of the underlying linear family

\item \code{darccirc_fr()}: circular density

\item \code{rarccirc_fr()}: random generation

}}
