% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\docType{package}
\name{arcstat-package}
\alias{arcstat}
\alias{arcstat-package}
\title{arcstat: Arc-Length Statistics: Goodness of Fit, Distributions and a Bayesian Test}
\description{
Inference from the arc length of statistical functions. Three tools share one pure-C back-end: a goodness-of-fit test based on the arc length of the probability plot, with an analytic saddlepoint null and sensitivity to local density structure that the empirical-distribution tests miss; two constructions that build a distribution from the arc length of its defining curve, the arc-length generator and the quantile arc-length family estimated by L-moments; and a Bayesian nonparametric arc-length goodness-of-fit test on the Dirichlet-process posterior. The same C sources back the 'Python' package 'arcstat'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mtloots/arcstat}
  \item Report bugs at \url{https://github.com/mtloots/arcstat/issues}
}

}
\author{
\strong{Maintainer}: M. Theodor Loots \email{theo.loots@gmail.com} (\href{https://orcid.org/0000-0001-7722-3913}{ORCID})

Authors:
\itemize{
  \item M. Theodor Loots \email{theo.loots@gmail.com} (\href{https://orcid.org/0000-0001-7722-3913}{ORCID})
}

}
\keyword{internal}
