% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{bc_cdf}
\alias{bc_cdf}
\title{Beta-companion distribution function}
\usage{
bc_cdf(x, alpha, beta)
}
\arguments{
\item{x}{quantiles at which to evaluate.}

\item{alpha, beta}{the two exponents of the quantile density.}
}
\value{
A numeric vector of probabilities, the same length as \code{x}.
}
\description{
The inverse of \code{\link{bc_q}}, obtained in the back-end by a continued-fraction
incomplete beta and a safeguarded root find.
}
\examples{
bc_cdf(bc_q(c(0.3, 0.7), -0.60, -0.35), -0.60, -0.35)
}
