% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{bc_pdf}
\alias{bc_pdf}
\title{Beta-companion density}
\usage{
bc_pdf(x, alpha, beta)
}
\arguments{
\item{x}{points at which to evaluate the density.}

\item{alpha, beta}{the two exponents of the quantile density.}
}
\value{
A numeric vector of density values, the same length as \code{x}.
}
\description{
The derivative of \code{\link{bc_cdf}}, equal to the reciprocal of the quantile density
evaluated at the corresponding probability.
}
\examples{
## the density is the derivative of the distribution function
xx <- bc_q(0.3, -0.60, -0.35); eps <- 1e-6
(bc_cdf(xx + eps, -0.60, -0.35) - bc_cdf(xx - eps, -0.60, -0.35)) / (2 * eps)
bc_pdf(xx, -0.60, -0.35)
}
