% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{k4_band_model}
\alias{k4_band_model}
\alias{k4_band_sample}
\title{Band arc lengths of a scaled kappa curve and of a data polyline}
\usage{
k4_band_model(theta, breaks, nodes = 60L)

k4_band_sample(x, y, breaks)
}
\arguments{
\item{theta}{c(g0, g1, mu, sg, k, h)}

\item{breaks}{band break points (length J+1)}

\item{nodes}{number of Gauss-Legendre nodes per band}

\item{x, y}{data ordered in x}
}
\value{
A numeric vector of band arc lengths, one per band.
}
\description{
Band arc lengths of a scaled kappa curve and of a data polyline
}
