% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{k4_q}
\alias{k4_q}
\alias{k4_cdf}
\alias{k4_pdf}
\title{Four-parameter kappa quantile, distribution and density functions}
\usage{
k4_q(u, mu = 0, sg = 1, k, h)

k4_cdf(x, mu = 0, sg = 1, k, h)

k4_pdf(x, mu = 0, sg = 1, k, h)
}
\arguments{
\item{u, x}{numeric vectors of probabilities or quantiles}

\item{mu, sg, k, h}{kappa parameters (location, scale, two shapes)}
}
\value{
numeric vector
}
\description{
Four-parameter kappa quantile, distribution and density functions
}
