% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{k4_tau34}
\alias{k4_tau34}
\title{Theoretical L-moment ratios of the standard kappa distribution}
\usage{
k4_tau34(k, h, nodes = 200L)
}
\arguments{
\item{k, h}{shape parameters}

\item{nodes}{number of Gauss-Legendre nodes}
}
\value{
c(tau3, tau4, l1, l2)
}
\description{
Theoretical L-moment ratios of the standard kappa distribution
}
