% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstat.R
\name{sample_lmoments}
\alias{sample_lmoments}
\title{Sample L-moments of data}
\usage{
sample_lmoments(x, nmom = 4L)
}
\arguments{
\item{x}{data.}

\item{nmom}{number of L-moments.}
}
\value{
A numeric vector of \code{nmom} sample L-moments.
}
\description{
Sample L-moments of data
}
\examples{
## for a uniform grid on (0,1) the first L-moment is 1/2 and the second is 1/6
u <- (seq_len(20000) - 0.5) / 20000
sample_lmoments(u, 4L)[1:2]
}
